requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
gunicorn==21.2.0
pillow==10.2.0
requests==2.31.0